feat: Refactor command-line argument parsing into a dedicated utility and introduce new logging and templates for instrument profile calibration. feat: Automated ipynb+pdf calibration report generation TODO: test calibrate.bat stability on Windows and incorporate script changes to github build workflow.
