+++ conf/logging_calibration.properties
@@ -3,7 +3,7 @@ handlers=java.util.logging.FileHandler
 java.util.logging.FileHandler.level=ALL
 java.util.logging.FileHandler.formatter=com.medals.libsdatagenerator.util.CustomLogFormatter
 java.util.logging.FileHandler.limit=3000000
-java.util.logging.FileHandler.pattern=__LOG_PATH_PLACEHOLDER__/LIBSDataGenerator%g.log
+java.util.logging.FileHandler.pattern=__LOG_PATH_PLACEHOLDER__/InstrumentProfileCalibration%g.log
 java.util.logging.FileHandler.count=10
 java.util.logging.FileHandler.append=true
