+++ src/test/java/com/medals/libsdatagenerator/controller/InstrumentProfileControllerTest.java
@@ -1,5 +1,6 @@
 package com.medals.libsdatagenerator.controller;
 
+import com.medals.libsdatagenerator.util.CmdlineParserUtil;
 import org.apache.commons.cli.CommandLine;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +11,7 @@ class InstrumentProfileControllerTest {
     @Test
     void testParseCommandLineArgsValid() {
         String[] args = {"-i", "data.csv", "-c", "Fe-100", "-o", "out.json", "-n", "TestInst"};
-        CommandLine cmd = InstrumentProfileController.parseCommandLineArgs(args);
+        CommandLine cmd = new CmdlineParserUtil().parseCommandLineArgsForCalibration(args);
         
         assertNotNull(cmd);
         assertEquals("data.csv", cmd.getOptionValue("i"));
@@ -24,14 +25,14 @@ class InstrumentProfileControllerTest {
         String[] args = {"-i", "data.csv"}; // Missing -c
         
         // Should catch ParseException inside and return null
-        CommandLine cmd = InstrumentProfileController.parseCommandLineArgs(args);
+        CommandLine cmd = new CmdlineParserUtil().parseCommandLineArgsForCalibration(args);
         assertNull(cmd);
     }
 
     @Test
     void testParseCommandLineArgsHelp() {
         String[] args = {"--help"};
-        CommandLine cmd = InstrumentProfileController.parseCommandLineArgs(args);
+        CommandLine cmd = new CmdlineParserUtil().parseCommandLineArgsForCalibration(args);
         assertNull(cmd); // Returns null for help
     }
 }
